Convert usize to u64 in Store trait APIs
APIs in `StoreLike` now uses `u64` instead of `usize` It's because
`u64` expresses the size of the data, but `usize` expresses local
memory limits.

+++ nativelink-config/src/stores.rs
@@ -763,12 +763,12 @@ pub struct S3Store {
     /// upload will be aborted and the client will likely receive an error.
     ///
     /// Default: 5MB.
-    pub max_retry_buffer_per_request: Option<usize>,
+    pub max_retry_buffer_per_request: Option<u64>,
 
     /// Maximum number of concurrent UploadPart requests per MultipartUpload.
     ///
     /// Default: 10.
-    pub multipart_max_concurrent_uploads: Option<usize>,
+    pub multipart_max_concurrent_uploads: Option<u64>,
 
     /// Allow unencrypted HTTP connections. Only use this for local testing.
     ///
